Bug fix for - Refactor File structure
